requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
ultralytics>=8.0.0  # YOLOv8
opencv-python>=4.5.0  # cv2
python-multipart>=0.0.5
numpy>=1.21.0
pytest>=7.0.0
pylint>=2.12.0
black>=22.0.0
python-dotenv>=0.19.0
mlflow>=2.0.0
redis>=4.0.0
minio>=7.0.0
pyyaml>=6.0.0  # yaml
torch>=2.0.0
albumentations>=1.3.0  # Advanced augmentations
scikit-image>=0.19.0  # Image processing
torchvision>=0.15.0  # Required for transforms
timm>=0.9.0  # For model ensembling
scipy>=1.9.0  # For advanced filtering
pillow>=9.0.0  # PIL
matplotlib>=3.5.0  # For visualization
seaborn>=0.12.0  # For enhanced plotting
pandas>=1.5.0  # For measurement analysis
tqdm>=4.65.0  # Progress bars
jsonlines>=3.1.0  # For measurement logging
opencv-contrib-python>=4.5.0  # Advanced OpenCV features
xgboost>=1.7.0
scikit-learn>=1.0.0